granite-coder/pyproject.toml
[project]
name = "granite-coder"
version = "0.1.0"
description = "Token-efficient coding agent using IBM Granite 4 and OneCoder tools"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=0.1.0",
    "ollama>=0.1.6",
    "pydantic>=2.0.0",
    "click>=8.0.0",
    "rlms>=0.1.0"
]
[project.scripts]
granite-coder = "src.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.metadata]
allow-direct-references = true
